% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_flextable}
\alias{as_flextable}
\title{Convert input object to flextable}
\usage{
as_flextable(x)
}
\arguments{
\item{x}{An object to convert to flextable}
}
\value{
A \code{\link[flextable:flextable]{flextable::flextable()}} object.
}
\description{
This function is short for the \code{\link[gtsummary:deprecated]{gtsummary::as_flextable()}}
function to provide cleaner code within the package.
}
\examples{
\dontrun{
as_flextable(gtsummary::tbl_summary(iris))
}
}
