% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_counts_at_or_above_threshold}
\alias{calculate_counts_at_or_above_threshold}
\title{Calculate counts at or above a threshold for variables}
\usage{
calculate_counts_at_or_above_threshold(x, vars, threshold = 0, by = NULL)
}
\arguments{
\item{x}{Data frame containing variables}

\item{vars}{List of variables to count number at or above threshold}

\item{threshold}{A string representing the threshold factor level.}

\item{by}{An optional grouping variable to arrange data by}
}
\value{
A data frame of counts (see details).
}
\description{
Calculate the counts of entries exceeding a threshold value
for a series of variables.
}
