% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_numerical_summaries}
\alias{calculate_numerical_summaries}
\title{Calculate numerical summaries of tibble for plotting}
\usage{
calculate_numerical_summaries(x, vars, by = NULL)
}
\arguments{
\item{x}{Data frame to summarize variables in}

\item{vars}{Variables to summarize}

\item{by}{Optionally group by a column of data prior to summarization}
}
\value{
A table of summarized variables from \code{x}, including the count, mean, sd, se.
}
\description{
Calculate standard statistics (mean, sd, se, n) from a
list of variables in a tibble.
}
\note{
At the moment, \code{by} is a string.
}
\examples{
\dontrun{
calculate_numerical_summaries(iris, vars=c("Sepal.Length","Sepal.Width"),by="Species")
}
}
