% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_categorical.R
\name{categorical}
\alias{categorical}
\title{Categorical data type}
\description{
A categorical type is a question involving one response out of
a set of options. These options can be treated as numeric if on a likert
scale, but generally are categorical.
}
\details{
With a categorical variable, the summarizations are frequency
tables by category level.

Possibilities include
\itemize{
\item summarize_categorical: Create a summarized table of the question(s)
\item plot_categorical: Create a figure of the categorical response frequencies.
\item gvsummary_categorical: Create a flextable and plot of the categorical responses.
}
}
