% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_categorical_followup.R
\name{categorical_followup}
\alias{categorical_followup}
\title{Categorical Followup data type}
\description{
A categorical followup type is a question involving one response out of
a set of options. These options can be treated as numeric if on a likert
scale, but generally are categorical. The followup is a question like "Other, please specify"
}
\details{
With a categorical variable, the summarizations are frequency
tables by category level.

Possibilities include
\itemize{
\item summarize_categorical_followup: Create a summarized table of the question(s) and append the
followup responses.
\item plot_categorical_followup: Create a figure of the categorical response frequencies (no
followup responses).
\item gvsummary_categorical_followup: Create a flextable and plot of the categorical responses.
}
}
