% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_choice.R
\name{choice}
\alias{choice}
\title{Choice data type}
\description{
A choice type is a "select all that apply" question. This translates into
multiple responses for each participant.
}
\details{
When counting these types of variables, one can
include the number of participants that responded with a particular choice and show
all possible choices.

Possibilities include
\itemize{
\item summarize_choice: Create a summarized table of the choices.
\item plot_choice: Create a figure of the choices.
\item gvsummary_choice: Create a flextable and plot of the choices.
}
}
