% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_choice_followup.R
\name{choice_followup}
\alias{choice_followup}
\title{Choice Followup data type}
\description{
A choice followup type is a "select all that apply" question and a "Other, please
specify" followup-type question. The choice translates into multiple responses for each participant
(see \link{choice}). The result is a summary choice table \code{\link[=summarize_choice]{summarize_choice()}} and a comments table
\code{\link[=summarize_comments]{summarize_comments()}} combined.
}
\details{
When counting choice types of variables, one can
include the number of participants that responded with a particular choice and show
all possible choices. See the \link{choice} page for more details on this type. Comments have
their own options to consider (see \link{comments}). This type combines the two.

Possibilities include
\itemize{
\item summarize_choice_followup: Create a summarized table of the choices and a followup question.
}
}
