% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{collapse_levels}
\alias{collapse_levels}
\title{Collapse Value/Label pairs in a tibble to a string.}
\usage{
collapse_levels(x)
}
\arguments{
\item{x}{A tibble representing a variable's levels (or empty).}
}
\value{
A string representing the values/labels for a variable.
}
\description{
The data dictionary provides a mapping of values to labels for a variable. This function
converts such a table to a string representing the mapping.
}
\details{
This function assumes as input a \code{Levels} tibble which can be converted into string representation.
}
\note{
To convert all levels in the dictionary, see \code{\link{format_levels}}. This function converts only the levels
for a single variable.
}
\examples{
\dontrun{
collapse_levels(tibble::tibble(Value=c(1,2,3), Label=c("One","Two","Three")))

}
}
