% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{combo_numeric}
\alias{combo_numeric}
\title{Title}
\usage{
combo_numeric(
  x,
  vars,
  dictionary,
  by = NULL,
  header = "**Characteristic**",
  summary = "mean",
  default_wrap_length = 35,
  threshold = 3,
  title = "",
  col = c("#4F81BD", "#FAAB18", "#868686FF", "#CD534CFF")
)
}
\arguments{
\item{col}{}
}
\description{
Title
}
