% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_length}
\alias{count_length}
\title{Count length of x}
\usage{
count_length(x, vars)
}
\arguments{
\item{x}{A data frame}

\item{vars}{A list of variables for counting (not used)}
}
\value{
A count of x
}
\description{
Count length of x
}
