% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_yes}
\alias{count_yes}
\title{Count the number of Yes values in a list}
\usage{
count_yes(x, na.rm = TRUE)
}
\arguments{
\item{x}{A list, assumed to be a factor variable with values \code{Yes} and \code{No}.}
}
\value{
A count of the number of values that are \code{Yes}
}
\description{
Count the number of Yes values in a list
}
