% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_labels}
\alias{default_labels}
\title{Retrieve labels from data frame}
\usage{
default_labels(x, vars = colnames(x))
}
\arguments{
\item{x}{A data frame, possibly with variable labels.}

\item{vars}{A list of variables to restrict label extraction to.}
}
\value{
A named list of variable labels. Names are the columns
in the data frame. Values are the labels (or column names if
labels are missing).
}
\description{
Retrieve variable labels from data frame columns,
using the colnames for labels that are not present.
}
\details{
The \code{\link[labelled:var_label]{labelled::var_label()}} function provides variable
labels within a data frame. However, if the variable label is
missing it returns NULL. This function provides a reasonable
default in the case of NULL, namely the column name itself.
}
\examples{
\dontrun{
default_labels(ToothGrowth, vars=c("supp"))
}
}
