% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{docx_add}
\alias{docx_add}
\title{Add content to word document}
\usage{
docx_add(
  doc,
  header = "",
  text = NULL,
  fig = NULL,
  ft = NULL,
  level = 1,
  pre_page_break = FALSE,
  post_page_break = FALSE,
  fig.height = 5,
  fig.width = 6
)
}
\arguments{
\item{doc}{Word document object.}

\item{header}{Header text to include in document.}

\item{text}{Any text to add to in the paragraph below the header.}

\item{level}{The Heading level to use (default 1).}
}
\value{
Word document object
}
\description{
Adds content to a word document (including header).
}
