% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{docx_add_flextable_section}
\alias{docx_add_flextable_section}
\title{Add flextable at Heading 1}
\usage{
docx_add_flextable_section(doc, header = "", text = "", ft)
}
\arguments{
\item{doc}{Word document object.}

\item{header}{Header text to include in document.}

\item{text}{Any text to add to in the paragraph below the header.}

\item{ft}{Flextable object to add to word document.}
}
\value{
Word document object
}
\description{
Add flextable at Heading 1
}
