% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{docx_add_section}
\alias{docx_add_section}
\title{Add section text to a word document}
\usage{
docx_add_section(doc, header = "", text = "", level = 1)
}
\arguments{
\item{doc}{Word document object.}

\item{header}{Header text to include in document.}

\item{text}{Any text to add to in the paragraph below the header.}

\item{level}{The Heading level to use (default 1).}
}
\value{
Word document object
}
\description{
Adds section text to a word document (including header).
}
