% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{docx_add_subsection}
\alias{docx_add_subsection}
\title{Add Heading 2 word content}
\usage{
docx_add_subsection(doc, header = "", text = "", ft)
}
\arguments{
\item{doc}{Word document object.}

\item{...}{Other parameters to \code{\link{docx_add}}}
}
\value{
Word document object
}
\description{
Add Heading 2 word content
}
