% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_choice.R
\name{extract_choice_components}
\alias{extract_choice_components}
\title{Extract choice components from a data frame}
\usage{
extract_choice_components(x, vars)
}
\arguments{
\item{x}{A data frame with choice variables included.}

\item{vars}{A list of variables (assumed to be choices) from which to
extract choice information.}
}
\value{
A list consisting of
\itemize{
\item vars: the column names of the choices fields.
\item question: the question asked (in common among choices).
\item choice: the specific choice corresponding to a \code{var}.
}
}
\description{
A choice consists of a series of indicator columns for a
variable. The label for each variable is the question asked followed
by the specific choice (the text is \code{(choice=zzzz)}). This function
extracts the variables names, choice text and question text into a list.
}
