% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_var_labels}
\alias{extract_var_labels}
\title{Extract variable labels as a tibble}
\usage{
extract_var_labels(x, vars = dplyr::everything())
}
\arguments{
\item{x}{A tibble to extract labels from.}

\item{vars}{A list of variables to extract labels (can use \code{\link[tidyselect:tidyselect-package]{tidyselect::tidyselect()}} syntax).}
}
\value{
A tibble representing the labels. Two columns are returned: \code{variable} and \code{label}.
}
\description{
Variable labels can be set on tibble column names. This function
extracts them and formats them as a tibble for easy manipulation.
}
\details{
Variable labels in the \code{labelled} package allow one to create labels that may be
question text or other longer-form text associated with a variable. The \code{labelled} package
provides a way to get at these, using the \code{\link[labelled]{var_label}} function. This
function returns a named list. However, within the tidyverse it's often easier to manipulate
the labels as a tibble. This function provides that.

The returned tibble has two columns:
\itemize{
\item variable
\item label
}
}
\examples{
extract_var_labels(labelled::set_variable_labels(tibble::tribble(~a,~b,1,2), a="First variable", b="Second variable"))
}
