% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_wrap}
\alias{fct_wrap}
\title{Wrap the text of factor levels}
\usage{
fct_wrap(x, width = 25)
}
\arguments{
\item{x}{A list of factor data}

\item{width}{Width to wrap text.}
}
\value{
A list of factor data with the levels wrapped at \code{width}.
}
\description{
Rewrite a factor by wrapping the levels of the factor
to a shorter width (per line). Will coerce a non-factor to factor
first, then shorten the labels.
}
\examples{
\dontrun{
fct_wrap(paste(iris$Species, "iris", sep=" "), width=2)
}
}
