% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_categorical.R
\name{flextable_categorical}
\alias{flextable_categorical}
\title{Generate a flextable of categorical variables in a data frame}
\usage{
flextable_categorical(
  x,
  vars,
  header = "**Characteristic**",
  by = NULL,
  include_overall = FALSE
)
}
\arguments{
\item{x}{The data frame to summarize}

\item{vars}{(default: dplyr::everything) Columns of the data frame
that contain categorical fields}

\item{header}{(default: "\strong{Characteristic}") The summary table header}

\item{by}{Variable to stratify summary by.}

\item{include_overall}{Include a column of overall summary. Note this only
works (and makes sense) when the \code{by} variable is used.}
}
\value{
A \code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary()}} object representing the categorical
breakdown of \code{vars} from \code{x}.
}
\description{
Generate a flextable of categorical variables in a data frame
}
\examples{
\dontrun{
flextable_categorical(ToothGrowth, vars=c("dose"),
   by = "supp", include_overall=TRUE)
}
}
