% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_comments.R
\name{flextable_comments}
\alias{flextable_comments}
\title{Generate a flextable of comments in a data frame}
\usage{
flextable_comments(x, vars = dplyr::everything(x), header = "**Comments**")
}
\arguments{
\item{x}{A data frame}

\item{vars}{(default: dplyr::everything()) Variables that
have comments}

\item{header}{(default: "\strong{Comments}") Header text of the
flextable}
}
\value{
A comment table formatted as a \code{\link[flextable:flextable]{flextable::flextable()}}
}
\description{
Generate a flextable of comments in a data frame
}
\examples{
\dontrun{
tibble::tibble("This is a question"=c("comment 1","comment 2")) |>
  flextable_comments()
}
}
