% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_followup.R
\name{flextable_followup}
\alias{flextable_followup}
\title{Title}
\usage{
flextable_followup(
  x,
  vars,
  followup,
  header = "",
  by = NULL,
  group_headers = c("", ""),
  followup_statistic = "{n}",
  include_overall = FALSE
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
