% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{force_tbl_caption}
\alias{force_tbl_caption}
\title{Force table caption}
\usage{
force_tbl_caption()
}
\description{
With flextables, captions do not print correctly. This forces
the caption to print.
}
\details{
When flextables are used in word documents, the captions from the
quarto document do not render correctly. See
(https://stackoverflow.com/questions/77167685/cross-references-and-captions-not-working-when-using-gt-in-quarto-with-word-outp)
for details. The solution is to generate an empty html \code{\link[knitr:kable]{knitr::kable()}} table
which acts as the table reference for quarto.
}
