% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{format_calculations_for_plotting}
\alias{format_calculations_for_plotting}
\title{Title}
\usage{
format_calculations_for_plotting(x, xorig, vars)
}
\arguments{
\item{vars}{}
}
\value{

}
\description{
Given
}
