% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{format_levels}
\alias{format_levels}
\title{Convert data dictionary levels to strings}
\usage{
format_levels(x)
}
\arguments{
\item{x}{A dictionary column representing levels}
}
\description{
A function to convert the data dictionary levels into string representations.
}
\details{
The data dictionary contains information about specific variables in a redcap dataset. This
routine will take as input the Levels variable of the dictionary and turn it into a vector of strings.

The idiom should be:

\if{html}{\out{<div class="sourceCode">}}\preformatted{dictionary \%>\% mutate(levels_as_string = format_levels(Levels))
}\if{html}{\out{</div>}}
}
