% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_choice.R
\name{get_choice_list}
\alias{get_choice_list}
\title{Retrieve a choice list from a variable stem}
\usage{
get_choice_list(x, stem)
}
\arguments{
\item{x}{A data frame}

\item{stem}{A variable stem, for which choices are represented as 'stem___1'
and 'stem___2', etc.}
}
\value{
A vector of column names in x that correspond to the choices identified
for the stem.
}
\description{
A list of choices are prepended with '___N' to indicate the
choice. This function finds the choices from the variable stem from a
data frame and returns the column names of the choices.
}
\examples{
get_choice_list(
 setNames(iris, paste("example",1:5,sep="___")),
 stem = "example"
)
}
