% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_choice.R
\name{get_choices}
\alias{get_choices}
\title{Get a list of choices from a data frame}
\usage{
get_choices(x, stem)
}
\arguments{
\item{x}{A data frame with choices.}

\item{stem}{The variable stem that will be used to identify choices. Choice
columns are named as \code{stem___1}, \code{stem___2}, etc.}
}
\value{
A list consisting of
\itemize{
\item vars: the column names of the choices fields.
\item question: the question asked (in common among choices).
\item choice: the specific choice corresponding to a \code{var}.
}
}
\description{
A choice set is a collection of columns of a data frame
that have a common question associated, but are indicators of the
specific choice made (in a select all that apply). This function
extracts the choices as a list of column names, choice values and
the overall question asked.
}
