% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{get_factor_levels}
\alias{get_factor_levels}
\title{Extract formatted factor levels from a dictionary for a list of variables.}
\usage{
get_factor_levels(dictionary, vars)
}
\arguments{
\item{dictionary}{Data dictionary}

\item{vars}{Variable list to extract factor levels}
}
\value{
A unique list of strings (ideally one) representing the levels.
}
\description{
Extract formatted factor levels from a dictionary for a list of variables.
}
