% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{get_variable_range}
\alias{get_variable_range}
\title{Title}
\usage{
get_variable_range(dictionary, vars)
}
\arguments{
\item{dictionary}{A data dictionary (table).}

\item{vars}{A list of variables to extract ranges for.}
}
\value{

}
\description{
Title
}
