% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvsummary.R
\name{gvsummary_categorical}
\alias{gvsummary_categorical}
\title{Title}
\usage{
gvsummary_categorical(
  x,
  vars = colnames(x),
  header = "**Characteristic**",
  by = NULL,
  include_overall = FALSE,
  dictionary = NULL,
  col = c("#4F81BD", "#FAAB18", "#868686FF", "#CD534CFF"),
  title = "",
  fit = "fixed",
  xlab = NULL,
  combine = FALSE
)
}
\arguments{
\item{xlab}{A label for the x axis (or vector of labels if multiple vars). Default
is the \code{\link[labelled:var_label]{labelled::var_label()}} or \code{vars} otherwise.}

\item{combine}{(FALSE) Should the table and plot be combined into a
single plot output?}
}
\value{
Depending on the value of \code{combine} either: a list consisting
of two elements: \code{plot} and \code{table}; or a single plot combining the
two elements.
}
\description{
Title
}
\examples{
\dontrun{
gvsummary_categorical(ToothGrowth, vars="dose",by="supp")
}
}
