% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvsummary.R
\name{gvsummary_categorical_followup}
\alias{gvsummary_categorical_followup}
\title{Generate paired table and graph for categorical data with followup}
\usage{
gvsummary_categorical_followup(
  x,
  vars = colnames(x),
  followup,
  followup_statistic = "{n}",
  by = NULL,
  group_headers = c("", ""),
  header = "**Characteristic**",
  include_overall = FALSE,
  dictionary = NULL,
  col = c("#4F81BD", "#FAAB18", "#868686FF", "#CD534CFF"),
  title = "",
  fit = "fixed",
  xlab = NULL,
  combine = FALSE
)
}
\arguments{
\item{combine}{}
}
\description{
This is a high-level function wrapping a number of different
activities: summarize categorical variables in a data frame and generate a
summary table and summary graph as a \code{\link[grid:grid.grob]{grid::grob()}}. A followup is free text
included as a followup question.

@details This function is a high-level summarization of a great deal of complex
processing code. As such, it is currently likely to be somewhat unstable. Briefly,
this function summarizes categaorical data elements within a data frame. Categorical
elements are listed by level. This function first builds a summary table of the
field(s). A followup question (free text) is included below the summarization
counts. Then a table of these summaries is generated. Additionally, the
fields are then graphed as bar plots. The table and figure are assembled into an
overall figure suitable for plotting. Note that this makes a very large
\code{\link[patchwork:patchwork-package]{patchwork::patchwork()}} grob so please take that into consideration when using
a markdown/officedown format.

@note The table output is always a \link[flextable:flextable]{flextable::flextable}.
}
