% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvsummary.R
\name{gvsummary_followup}
\alias{gvsummary_followup}
\title{Title}
\usage{
gvsummary_followup(
  x,
  vars = colnames(x),
  followup,
  followup_statistic = "{n}",
  by = NULL,
  group_headers = c("", ""),
  header = "**Characteristic**",
  include_overall = FALSE,
  dictionary = NULL,
  col = c("#4F81BD", "#FAAB18", "#868686FF", "#CD534CFF"),
  title = "",
  fit = "fixed",
  xlab = NULL,
  combine = FALSE
)
}
\arguments{
\item{combine}{}
}
\description{
Title
}
