% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvsummary.R
\name{gvsummary_numeric}
\alias{gvsummary_numeric}
\title{Generate paired table and graph for numeric data}
\usage{
gvsummary_numeric(
  x,
  vars,
  dictionary = NULL,
  by = NULL,
  header = "**Characteristic**",
  summary = "mean",
  wrap_length = 35,
  threshold = 3,
  title = "",
  col = c("#4F81BD", "#FAAB18", "#868686FF", "#CD534CFF"),
  fit = "fixed",
  xlab = NULL,
  combine = FALSE
)
}
\arguments{
\item{x}{A data frame to summarize}

\item{vars}{Variable (column) names in table}

\item{col}{}
}
\description{
This is a high-level function wrapping a number of different
activities: summarize numeric variables in a data frame and generate a
summary table and summary graph as a \code{\link[grid:grid.grob]{grid::grob()}}.

@details This function is a high-level summarization of a great deal of complex
processing code. As such, it is currently likely to be somewhat unstable. Briefly,
this function summarizes numeric data elements within a data frame. Numeric data
elements can be averaged. This function first builds a summary table of the numeric
field(s). Then a table of these summaries is generated. Additionally, the numeric
fields are then graphed as bar plots. The table and figure are assembled into an
overall figure suitable for plotting. Note that this makes a very large
\code{\link[patchwork:patchwork-package]{patchwork::patchwork()}} grob so please take that into consideration when using
a markdown/officedown format.
}
\examples{
\dontrun{
gvsummary_numeric(iris, vars=c("Sepal.Length","Sepal.Width"), by="Species")
}
}
