% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{levels_as_string}
\alias{levels_as_string}
\title{Extract levels of variables as a string}
\usage{
levels_as_string(x, vars = colnames(x))
}
\arguments{
\item{x}{A data frame of factors}

\item{vars}{A list of variables to choose from data frame}
}
\value{
A vector of strings where each string represents a concatenated
set of factor levels.
}
\description{
Extract from a data frame a subset of variables, then
create single-string descriptions of the levels of each of these
variables.
}
\examples{
\dontrun{
levels_as_string(iris, vars="Species")
}
}
