% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_numeric.R
\name{numeric}
\alias{numeric}
\title{Numeric data type}
\description{
A numeric type is a question involving a numeric output. These
can be options that can be treated as numeric if on a likert
scale, but can also be general numeric values.
}
\details{
With a numeric variable, the summarizations are mean/median, etc

Possibilities include
\itemize{
\item \code{\link[=summarize_numeric]{summarize_numeric()}}: Create a summarized table of the question(s)
\item \code{\link[=plot_numeric]{plot_numeric()}}: Create a figure of the numeric responses
\item \code{\link[=gvsummary_numeric]{gvsummary_numeric()}}: Create a flextable and plot of the numeric responses.
}
}
