% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variables.R
\name{plot_numeric}
\alias{plot_numeric}
\title{Plot numerical variables in a data frame}
\usage{
plot_numeric(
  x,
  vars,
  by = NULL,
  dictionary = NULL,
  wrap_length = 35,
  threshold = 3,
  title = "",
  col = c("#4F81BD", "#FAAB18", "#868686FF", "#CD534CFF"),
  xlab = NULL
)
}
\arguments{
\item{x}{A data frame to plot variable(s) from}

\item{by}{}
}
\description{
Plot numerical variables in a data frame
}
