% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variables.R
\name{plot_variables_as_yesno}
\alias{plot_variables_as_yesno}
\title{Plot variable counts}
\usage{
plot_variables_as_yesno(
  x,
  dictionary,
  by = NULL,
  default_wrap_length = 35,
  threshold = "",
  title = "",
  col = c("#4F81BD", "#FAAB18", "#868686FF", "#CD534CFF")
)
}
\arguments{
\item{x}{A data frame to plot variable(s) from}

\item{title}{A title for the plot}

\item{col}{Colors for the plot}
}
\description{
Plot variable counts
}
