% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variables.R
\name{plot_yesno}
\alias{plot_yesno}
\title{Plot yes/no variable frequency of yes}
\usage{
plot_yesno(
  x,
  vars,
  by = NULL,
  wrap_length = 35,
  xlab = NULL,
  title = "",
  col = c("#4F81BD", "#FAAB18", "#868686FF", "#CD534CFF")
)
}
\arguments{
\item{x}{A data frame to plot variable(s) from}

\item{vars}{A list of column names to plot}

\item{by}{A grouping column}

\item{wrap_length}{Length of factor level to wrap strings at}

\item{title}{A title for the plot}

\item{col}{Colors for the plot}
}
\description{
Plot yes/no variable frequency of yes
}
