% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_and_na}
\alias{select_and_na}
\title{Reduce data frame and set empty strings to NA}
\usage{
select_and_na(x, vars)
}
\arguments{
\item{x}{A data frame}

\item{vars}{The variables to select}
}
\description{
Reduce data frame and set empty strings to NA
}
