% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_basic.R
\name{style_basic_as_flextable}
\alias{style_basic_as_flextable}
\title{Convert basic summary table to flextable output}
\usage{
style_basic_as_flextable(x)
}
\arguments{
\item{x}{A summary table}
}
\value{
A flextable
}
\description{
The basic summary table can be converted to flextable output, with
a few simple styling options.
}
