% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_categorical.R
\name{style_categorical_flextable}
\alias{style_categorical_flextable}
\title{Apply style to catgorical flextable}
\usage{
style_categorical_flextable(x)
}
\arguments{
\item{x}{A \code{\link[flextable:flextable]{flextable::flextable()}} object for styling.}
}
\value{
A \code{\link[flextable:flextable]{flextable::flextable()}} object representing the categorical
table \code{x}, appropriately styled.
}
\description{
Apply a consistent style to a \code{\link[flextable:flextable]{flextable::flextable()}}
object appropriate for categorical output.
}
\details{
A \code{\link[flextable:flextable]{flextable::flextable()}} is useful visualiation, particularly within
a MS-Word document or HTML output. This function does appropriate
styling of a categorical flextable. Along the way, it applies some styling to the table
that has generally been helpful for visualization.
}
\examples{
\dontrun{
  summarize_categorical(ToothGrowth, vars=c("dose"), by="supp") |>
  as_flextable() |>
  style_categorical_flextable()

}
}
