% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_categorical_followup.R
\name{style_categorical_followup_flextable}
\alias{style_categorical_followup_flextable}
\title{Style flextable as a followup table}
\usage{
style_categorical_followup_flextable(x)
}
\arguments{
\item{x}{A flextable}
}
\value{
A flextable with specific styling applied
}
\description{
Apply consistent styling to a follow up flextable
to fit the typical contents of the table.
}
\examples{
\dontrun{
gtsummary::tbl_summary(iris) |>
  as_flextable() |>
  style_followup_flextable()
}
}
