% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_comments.R
\name{style_comments_as_flextable}
\alias{style_comments_as_flextable}
\title{Convert comment table to flextable output}
\usage{
style_comments_as_flextable(x)
}
\arguments{
\item{x}{}
}
\description{
Convert comment table to flextable output
}
