% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dichotomous.R
\name{style_dichotomous_flextable}
\alias{style_dichotomous_flextable}
\title{Apply style to catgorical flextable}
\usage{
style_dichotomous_flextable(x)
}
\arguments{
\item{x}{A \code{\link[flextable:flextable]{flextable::flextable()}} object for styling.}
}
\value{
A \code{\link[flextable:flextable]{flextable::flextable()}} object representing the dichotomous
table \code{x}, appropriately styled.
}
\description{
Apply a consistent style to a \code{\link[flextable:flextable]{flextable::flextable()}}
object appropriate for dichotomous output.
}
\details{
A \code{\link[flextable:flextable]{flextable::flextable()}} is useful visualiation, particularly within
a MS-Word document or HTML output. This function does appropriate
styling of a dichotomous flextable. Along the way, it applies some styling to the table
that has generally been helpful for visualization.
}
\examples{
\dontrun{
  summarize_dichotomous(ToothGrowth, vars=c("dose"), by="supp") |>
  as_flextable() |>
  style_dichotomous_flextable()

}
}
