% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style_flextable}
\alias{style_flextable}
\title{Default styling for a flextable}
\usage{
style_flextable(ft)
}
\arguments{
\item{ft}{A flextable to style.}
}
\value{
A flextable, with stylign applied.
}
\description{
This is the package default for styling a flextable. This adds color in the
header and uses the booktabs theme.
}
