% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_followup.R
\name{style_followup_as_flextable}
\alias{style_followup_as_flextable}
\title{Convert followup table to flextable output}
\usage{
style_followup_as_flextable(x)
}
\arguments{
\item{x}{A gtsummary table}
}
\value{
A flextable object representing x.
}
\description{
Convert followup table to flextable output
}
