% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_numeric.R
\name{style_numeric_flextable}
\alias{style_numeric_flextable}
\title{Style a numeric summary as a flextable}
\usage{
style_numeric_flextable(...)
}
\description{
Given a numeric summary, style the summary as a
\code{\link[flextable:flextable]{flextable::flextable()}}.
}
