% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_paired.R
\name{style_paired_as_flextable}
\alias{style_paired_as_flextable}
\title{Convert comment table to flextable output}
\usage{
style_paired_as_flextable(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Convert comment table to flextable output
}
