% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_basic.R
\name{summarize_basic}
\alias{summarize_basic}
\title{Create a basic summary table}
\usage{
summarize_basic(x, var, header = "**Characteristic**", by = NULL)
}
\arguments{
\item{by}{Variable to stratify summary by.}
}
\value{

}
\description{
The simplest summary table would be evaluating all variables as categorical. This
function generates this summary to remain consistent with more complex summary styles.
}
