% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_basic_numeric.R
\name{summarize_basic_numeric}
\alias{summarize_basic_numeric}
\title{Basic numerical summary table}
\usage{
summarize_basic_numeric(
  x,
  vars,
  dictionary,
  header = "**Characteristic**",
  summary = "mean",
  by = NULL
)
}
\arguments{
\item{dictionary}{The data dictionary for the instrument}

\item{summary}{}
}
\description{
Basic numerical summary table
}
