% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_categorical.R
\name{summarize_categorical}
\alias{summarize_categorical}
\title{Create a basic categorical variable breakdown table}
\usage{
summarize_categorical(
  x,
  vars,
  header = "**Characteristic**",
  by = NULL,
  include_overall = FALSE
)
}
\arguments{
\item{x}{The data frame to summarize}

\item{vars}{(default: dplyr::everything) Columns of the data frame
that contain categorical fields}

\item{header}{(default: "\strong{Characteristic}") The summary table header}

\item{by}{Variable to stratify summary by.}

\item{include_overall}{Include a column of overall summary. Note this only
works (and makes sense) when the \code{by} variable is used.}
}
\value{
A \code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary()}} object representing the categorical
breakdown of \code{vars} from \code{x}.
}
\description{
Extract variables from the table and format the variables
by value within a summary table.
}
\details{
This function combines (potentially) multiple columns of a
data frame into a single output summary table. This table has the
variable name (or column name if there is no \code{\link[labelled:var_label]{labelled::var_label()}})
followed by a summarization of levels for this variable. If your
variable is numeric, please see \code{\link[=summarize_numeric]{summarize_numeric()}}.
}
\examples{
\dontrun{
summarize_categorical(ToothGrowth, vars =c("dose"),by="supp")
}

}
\seealso{
style_categorical_flextable()
}
