% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_categorical_followup.R
\name{summarize_categorical_followup}
\alias{summarize_categorical_followup}
\title{Create a categorical table with followup comments}
\usage{
summarize_categorical_followup(
  x,
  vars,
  by = NULL,
  followup,
  followup_statistic = "{n}",
  header = "**Characteristic**",
  group_headers = c("", ""),
  include_overall = FALSE,
  output = c("gtsummary", "flextable", "rmarkdown")
)
}
\arguments{
\item{x}{Labelled data frame containing all data in survey}

\item{vars}{(default: dplyr::everything) Columns of the data frame
that contain categorical fields}

\item{by}{Variable to stratify summary by.}

\item{followup}{A column to use as the follow up question/response (comment).}

\item{followup_statistic}{A statistic (see \code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary()}}). Default
is \code{"{n}"}, meaning include counts.}

\item{header}{The overall header for the table}

\item{group_headers}{A vector of strings of header names for the questions (default
QUESTION) and followup (default FOLLOWUP).}

\item{include_overall}{(TRUE) Include a column of overall counts. Only useful when
stratifying (\code{by}).}

\item{output}{A string representing the desired output. Choices include:
\itemize{
\item "gtsummary" A \link[gtsummary:gtsummary-package]{gtsummary::gtsummary} object.
\item "flextable" A \link[flextable:flextable]{flextable::flextable} table with formatting applied to a gtsummary table.
\item "rmarkdown" Raw flextable output for use in rmarkdown (see \code{\link[flextable:flextable_to_rmd]{flextable::flextable_to_rmd()}}
for details). This is most suitable for running summarize in a loop or as list component.
}}
}
\value{
A table representing the frequency of \code{vars}, with
any additional free text responses in \code{followup} listed below the table.
}
\description{
Create a categorical table with followup comments
}
