% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_choice.R
\name{summarize_choice}
\alias{summarize_choice}
\title{Summarize choice variables to a table}
\usage{
summarize_choice(
  x,
  vars = NULL,
  by = NULL,
  include_overall = TRUE,
  output = c("gtsummary", "flextable", "rmarkdown")
)
}
\arguments{
\item{x}{A data frame}

\item{vars}{A list of variables (currently only one) to select. Note that as a
choice variable, this corresponds to a choice set "stem", or the text preceding
the \verb{___N} in the name.}

\item{by}{An optional stratification of the choices when summarizing.}

\item{include_overall}{(TRUE) Include a column of overall counts. Only useful when
stratifying (\code{by}).}

\item{output}{A string representing the desired output. Choices include:
\itemize{
\item "gtsummary" A \link[gtsummary:gtsummary-package]{gtsummary::gtsummary} object.
\item "flextable" A \link[flextable:flextable]{flextable::flextable} table with formatting applied to a gtsummary table.
\item "rmarkdown" Raw flextable output for use in rmarkdown (see \code{\link[flextable:flextable_to_rmd]{flextable::flextable_to_rmd()}}
for details). This is most suitable for running summarize in a loop or as list component.
}}
}
\value{
A summary table object which summarizes the counts of observations in
the choice variable(s). Results are in the format described in \code{output}.
}
\description{
Summarize choice variables to a table
}
