% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dichotomous.R
\name{summarize_dichotomous}
\alias{summarize_dichotomous}
\alias{summarize_dichotomous_as_flextable}
\alias{summarize_dichotomous_as_gtsummary}
\alias{summarize_dichotomous_as_rmarkdown}
\title{Create a basic dichotomous variable breakdown table}
\usage{
summarize_dichotomous(
  x,
  vars,
  header = "",
  by = NULL,
  include_overall = FALSE,
  output = c("gtsummary", "flextable", "rmarkdown")
)

summarize_dichotomous_as_flextable(...)

summarize_dichotomous_as_gtsummary(...)

summarize_dichotomous_as_rmarkdown(...)
}
\arguments{
\item{x}{The data frame to summarize}

\item{vars}{(default: dplyr::everything) Columns of the data frame
that contain dichotomous fields}

\item{header}{(default: "\strong{Characteristic}") The summary table header}

\item{by}{Variable to stratify summary by.}

\item{include_overall}{Include a column of overall summary. Note this only
works (and makes sense) when the \code{by} variable is used.}

\item{output}{A string representing the desired output. Choices include:
\itemize{
\item "gtsummary" A \link[gtsummary:gtsummary-package]{gtsummary::gtsummary} object.
\item "flextable" A \link[flextable:flextable]{flextable::flextable} table with formatting applied to a gtsummary table.
\item "rmarkdown" Raw flextable output for use in rmarkdown (see \code{\link[flextable:flextable_to_rmd]{flextable::flextable_to_rmd()}}
for details). This is most suitable for running summarize in a loop or as list component.
}}
}
\value{
A \code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary()}} object representing the dichotomous
breakdown of \code{vars} from \code{x}.
}
\description{
Extract variables from the table and format the variables
by value within a summary table.
}
\details{
This function combines (potentially) multiple columns of a
data frame into a single output summary table. This table has the
variable name (or column name if there is no \code{\link[labelled:var_label]{labelled::var_label()}})
followed by a summarization of levels for this variable. If your
variable is numeric, please see \code{\link[=summarize_numeric]{summarize_numeric()}}.
}
\section{Functions}{
\itemize{
\item \code{summarize_dichotomous_as_flextable()}: Flextable output of summarized dichotomous table

\item \code{summarize_dichotomous_as_gtsummary()}: gtsummary output of summarized dichotomous table

\item \code{summarize_dichotomous_as_rmarkdown()}: rmarkdown output of summarized dichotomous table

}}
\examples{
\dontrun{
summarize_dichotomous(ToothGrowth, vars =c("dose"),by="supp")
}

}
\seealso{
style_dichotomous_flextable()
}
