% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_followup.R
\name{summarize_followup}
\alias{summarize_followup}
\title{Create a categorical table with followup comments}
\usage{
summarize_followup(x, var, followup, header = "**Characteristic**")
}
\arguments{
\item{header}{}
}
\value{

}
\description{
Create a categorical table with followup comments
}
