% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_numeric.R
\name{summarize_numeric}
\alias{summarize_numeric}
\alias{summarize_numeric_as_flextable}
\alias{summarize_numeric_as_gtsummary}
\alias{summarize_numeric_as_rmarkdown}
\title{Summarize numeric variables}
\usage{
summarize_numeric(
  x,
  vars = colnames(x),
  by = NULL,
  dictionary = NULL,
  header = "**Characteristic**",
  summary = "mean",
  include_overall = TRUE,
  output = c("gtsummary", "flextable", "rmarkdown")
)

summarize_numeric_as_flextable(...)

summarize_numeric_as_gtsummary(...)

summarize_numeric_as_rmarkdown(...)
}
\arguments{
\item{x}{A data frame to summarize}

\item{vars}{Specific variables to summarize (default is all).}

\item{by}{Variable to stratify summary by.}

\item{dictionary}{The data dictionary for the instrument (optional)}

\item{header}{(default: "\strong{Characteristic}") The summary table header}

\item{summary}{A summary measure (default = "mean")}

\item{include_overall}{Include a column of overall summary. Note this only
works (and makes sense) when the \code{by} variable is used.}

\item{output}{A string representing the desired output. Choices include:
\itemize{
\item "gtsummary" A \link[gtsummary:gtsummary-package]{gtsummary::gtsummary} object.
\item "flextable" A \link[flextable:flextable]{flextable::flextable} table with formatting applied to a gtsummary table.
\item "rmarkdown" Raw flextable output for use in rmarkdown (see \code{\link[flextable:flextable_to_rmd]{flextable::flextable_to_rmd()}}
for details). This is most suitable for running summarize in a loop or as list component.
}}
}
\value{
A table object representing the numeric
breakdown of \code{vars} from \code{x}.
}
\description{
Summarize numeric variables
}
\section{Functions}{
\itemize{
\item \code{summarize_numeric_as_flextable()}: Flextable output of summarized numeric table

\item \code{summarize_numeric_as_gtsummary()}: gtsummary output of summarized numeric table

\item \code{summarize_numeric_as_rmarkdown()}: rmarkdown output of summarized numeric table

}}
\examples{
\dontrun{
summarize_numeric(iris, vars=c("Petal.Length","Petal.Width"))
}
}
