% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_paired.R
\name{summarize_paired}
\alias{summarize_paired}
\title{Title}
\usage{
summarize_paired(
  x,
  vars,
  dictionary,
  header = "**Characteristic**",
  summarizer = status ~ post - pre
)
}
\arguments{
\item{summarizer}{}
}
\value{

}
\description{
Title
}
