% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variables.R
\name{threshold_plot}
\alias{threshold_plot}
\title{Title}
\usage{
threshold_plot(x, vars, by = NULL, threshold, ...)
}
\arguments{
\item{x}{A data frame to plot variable(s) from}

\item{...}{}
}
\description{
Title
}
