% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{tidy_data_for_calculation}
\alias{tidy_data_for_calculation}
\title{Tidy data for summarization (Deprecated)}
\usage{
tidy_data_for_calculation(x, vars, by)
}
\arguments{
\item{by}{}
}
\description{
Prior to summarization, the data must be tidied for easier
summarization operations.
}
\details{
Most of the tidyverse works better with tidy data. This extends to the notion
of long data (vs wide data). This function will transform a data table to
a series of entries:
\itemize{
\item variable
\item value
\item group_by_field
}

After the transformation, the data is grouped by \code{variable} and \code{by} so that
summarization happens naturally via grouping.
}
\note{
The \code{by} variable can be NULL, in which case grouping occurs just by
variable.
}
