% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_numerical_summaries}
\alias{calculate_numerical_summaries}
\title{Calculate numerical summaries of tibble for plotting}
\usage{
calculate_numerical_summaries(x, vars, by = NULL)
}
\arguments{
\item{x}{Tibble to summarize}

\item{vars}{Variables to summarize}

\item{by}{}
}
\value{

}
\description{
Calculate standard statistics (mean, sd, se, n) from a
list of variables in a tibble.
}
\note{
At the moment, \code{by} is a bare word not a string.
}
