% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_wrap}
\alias{fct_wrap}
\title{Wrap the text of a factor, recreating the factor.}
\usage{
fct_wrap(x, width = 25)
}
\arguments{
\item{x}{A list of factor data}

\item{width}{Width to wrap text.}
}
\value{

}
\description{
Wrap the text of a factor, recreating the factor.
}
