% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{get_factor_levels}
\alias{get_factor_levels}
\title{Extract formatted factor levels from a dictionary for a
list of variables.}
\usage{
get_factor_levels(x, vars = colnames(x), dictionary = NULL)
}
\arguments{
\item{vars}{Variable list to extract factor levels}

\item{dictionary}{Data dictionary}
}
\value{
A unique list of strings (ideally one) representing the levels.
}
\description{
Extract formatted factor levels from a dictionary for a
list of variables.
}
