% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{nl}
\alias{nl}
\title{Short-hand add new line in word document}
\usage{
nl(doc)
}
\arguments{
\item{doc}{A officer word document}
}
\value{
A document with a blank paragraph added
}
\description{
Short-hand add new line in word document
}
