% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_comments.R
\name{style_comments_as_flextable}
\alias{style_comments_as_flextable}
\title{Style flextable with sensible defaults}
\usage{
style_comments_as_flextable(x, widths = c(5, 1))
}
\arguments{
\item{x}{A \code{\link[flextable:flextable]{flextable::flextable()}} object for styling.}

\item{widths}{(Default: c(5,1)): The relative widths of the
two columns representing comments (usually Question, Response).}
}
\value{
A \code{\link[flextable:flextable]{flextable::flextable()}} object representing the comment
table \code{x}, appropriately styled.
}
\description{
Style a \code{\link[flextable:flextable]{flextable::flextable()}} comments
object, applying default styling appropriate for comments.
}
\examples{
\dontrun{
 tibble::tibble("This is a question"=c("comment 1","comment 2")) |>
  summarize_comments() |>
  style_comments_as_flextable()

}
}
