% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_comments.R
\name{summarize_comments}
\alias{summarize_comments}
\title{Create a comment table}
\usage{
summarize_comments(x, vars, header = "**Characteristic**")
}
\arguments{
\item{vars}{}
}
\description{
Create a comment table
}
