% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_numeric.R
\name{summarize_numeric}
\alias{summarize_numeric}
\title{Summarize numeric variables}
\usage{
summarize_numeric(
  x,
  vars = colnames(x),
  dictionary = NULL,
  header = "**Characteristic**",
  summary = "mean",
  by = NULL
)
}
\arguments{
\item{x}{A data frame to summarize}

\item{vars}{Specific variables to summarize (default is all).}

\item{dictionary}{The data dictionary for the instrument (optional)}

\item{summary}{}
}
\description{
Summarize numeric variables
}
\examples{
\dontrun{
summarize_numeric(iris, vars=c("Petal.Length","Petal.Width"))
}
}
